+++ sdl2/test/sdl2ext_font_test.py
@@ -1,10 +1,11 @@
 # -*- coding: utf-8 -*-
 import pytest
+import sdl2
 from sdl2 import ext as sdl2ext
 from sdl2.ext.compat import byteify
 from sdl2.ext.pixelaccess import pixels2d
 from sdl2.ext.surface import _create_surface
-from sdl2 import surface, pixels
+from sdl2 import surface, pixels, SDL_ClearError, SDL_GetError
 
 _HASSDLTTF = True
 try:
@@ -14,6 +15,7 @@ except ImportError:
 
 
 RESOURCES = sdl2ext.Resources(__file__, "resources")
+RED_RGBA = (255, 0, 0, 255)
 FONTMAP = ["0123456789",
            "ABCDEFGHIJ",
            "KLMNOPQRST",
@@ -25,7 +27,25 @@ FONTMAP = ["0123456789",
            ]
 
 
-class TestSDL2ExtFont(object):
+@pytest.fixture(scope="module")
+def with_sdl():
+    sdl2ext.init()
+    yield
+    sdl2ext.quit()
+
+@pytest.fixture()
+def with_font_ttf(with_sdl):
+    if _HASSDLTTF:
+        SDL_ClearError()
+        fontpath = RESOURCES.get_path("tuffy.ttf")
+        font = sdl2ext.FontTTF(fontpath, 20, RED_RGBA)
+        assert SDL_GetError() == b""
+        assert font
+        yield font
+        font.close()
+
+
+class TestBitmapFont(object):
     __tags__ = ["sdl", "sdl2ext"]
 
     @classmethod
@@ -39,7 +59,7 @@ class TestSDL2ExtFont(object):
     def teardown_class(cls):
         sdl2ext.quit()
 
-    def test_BitmapFont(self):
+    def test_init(self):
         # Initialize surface and sprite for tests
         fontpath = byteify(RESOURCES.get_path("font.bmp"), "utf-8")
         sf = surface.SDL_LoadBMP(fontpath)
@@ -69,7 +89,7 @@ class TestSDL2ExtFont(object):
         with pytest.raises(TypeError):
             sdl2ext.BitmapFont([], (32, 32), FONTMAP)
 
-    def test_BitmapFont_render(self):
+    def test_render(self):
         # Initialize font and BitmapFont for tests
         fontpath = byteify(RESOURCES.get_path("font.bmp"), "utf-8")
         sf = surface.SDL_LoadBMP(fontpath)
@@ -85,7 +105,7 @@ class TestSDL2ExtFont(object):
         with pytest.raises(ValueError):
             font.render("this_should_fail")
 
-    def test_BitmapFont_render_text(self):
+    def test_render_text(self):
         # Initialize BitmapFont and dummy RGB888 surface for tests
         fontpath = byteify(RESOURCES.get_path("font.bmp"), "utf-8")
         font = sdl2ext.BitmapFont(fontpath)
@@ -135,7 +155,7 @@ class TestSDL2ExtFont(object):
         with pytest.raises(ValueError):
             font.render_text("this_should_fail")
 
-    def test_BitmapFont_render_on(self):
+    def test_render_on(self):
         np = pytest.importorskip("numpy", reason="numpy module is not available")
         # Initialize BitmapFont for tests
         fontpath = byteify(RESOURCES.get_path("font.bmp"), "utf-8")
@@ -164,7 +184,7 @@ class TestSDL2ExtFont(object):
         with pytest.raises(ValueError):
             font.render_on(target, "%nope")
 
-    def test_BitmapFont_remap(self):
+    def test_remap(self):
         # Initialize BitmapFont for tests
         fontpath = byteify(RESOURCES.get_path("font.bmp"), "utf-8")
         font = sdl2ext.BitmapFont(fontpath)
@@ -186,7 +206,7 @@ class TestSDL2ExtFont(object):
         with pytest.raises(ValueError):
             font.remap("h", 32, 1000, 32, 32)
 
-    def test_BitmapFont_contains(self):
+    def test_contains(self):
         # Initialize BitmapFont for tests
         fontpath = byteify(RESOURCES.get_path("font.bmp"), "utf-8")
         font = sdl2ext.BitmapFont(fontpath)
@@ -194,10 +214,10 @@ class TestSDL2ExtFont(object):
         for ch in "abcde12345.-,+":
             assert font.contains(ch)
         del font.offsets[" "]
-        for ch in " äöüß":
+        for ch in u" äöüß":
             assert not font.contains(ch)
 
-    def test_BitmapFont_can_render(self):
+    def test_can_render(self):
         sf = surface.SDL_LoadBMP(byteify(RESOURCES.get_path("font.bmp"),
                                          "utf-8"))
         assert isinstance(sf.contents, surface.SDL_Surface)
@@ -208,8 +228,220 @@ class TestSDL2ExtFont(object):
         assert font.can_render("473285435hfsjadfhriuewtrhefd")
         assert not font.can_render("testä")
 
-    @pytest.mark.skipif(not _HASSDLTTF, reason="SDL_TTF library not available")
-    def test_FontManager(self):
+
+
+@pytest.mark.skipif(not _HASSDLTTF, reason="SDL_TTF library not available")
+class TestFontTTF(object):
+    __tags__ = ["sdl", "sdl2ext"]
+
+    def test_init(self, with_sdl):
+        # Try opening and closing a font
+        fontpath = RESOURCES.get_path("tuffy.ttf")
+        font = sdl2ext.FontTTF(fontpath, 20, RED_RGBA)
+        assert SDL_GetError() == b""
+        assert font
+        font.close()
+
+        # Try opening a font with size specified in pt
+        font = sdl2ext.FontTTF(fontpath, "20pt", RED_RGBA)
+        assert SDL_GetError() == b""
+        assert font
+        font.close()
+
+        # Try opening a font with size specified in pixels
+        font = sdl2ext.FontTTF(fontpath, "20px", RED_RGBA)
+        assert SDL_GetError() == b""
+        assert font
+        
+        # Try opening a font with a custom set of height chars
+        chars = "aeiou"
+        font2 = sdl2ext.FontTTF(fontpath, "20px", RED_RGBA, height_chars=chars)
+        assert SDL_GetError() == b""
+        assert font
+        assert font._parse_size("20px") != font2._parse_size("20px")
+        font.close()
+        font2.close()
+
+        # Test exception on invalid font path
+        with pytest.raises(IOError):
+            sdl2ext.FontTTF("missing.ttf", 16, RED_RGBA)
+        
+        # Text exception on invalid size
+        with pytest.raises(ValueError):
+            sdl2ext.FontTTF(fontpath, 0, RED_RGBA)
+
+        # Test exception on missing height chars
+        with pytest.raises(RuntimeError):
+            sdl2ext.FontTTF(fontpath, 20, RED_RGBA, height_chars=u"的是不")
+
+    def test_get_ttf_font(self, with_font_ttf):
+        font = with_font_ttf
+        ttf_font = font.get_ttf_font()
+        assert isinstance(ttf_font, sdlttf.TTF_Font)
+
+        # Test exception for missing style
+        with pytest.raises(ValueError):
+            font.get_ttf_font("small")
+
+    def test_add_style(self, with_font_ttf):
+        font = with_font_ttf
+
+        # Add some new styles
+        font.add_style('large', '40px', RED_RGBA)
+        font.add_style('red', 20, pixels.SDL_Color(255, 0, 0))
+        font.add_style('red_on_white', 20, (255, 0, 0), (255, 255, 255))
+
+        # Test exception for existing style name
+        with pytest.raises(ValueError):
+            font.add_style('large', '50px', RED_RGBA)
+
+        # Test exception for invalid size unit
+        with pytest.raises(ValueError):
+            font.add_style('unit_err', '10cm', RED_RGBA)
+
+        # Test exception for non-integer size
+        with pytest.raises(ValueError):
+            font.add_style('float_err', '10.5pt', RED_RGBA)
+
+    def test_render_text(self, with_font_ttf):
+        font = with_font_ttf
+
+        # Try rendering some text
+        msg = "hello there!"
+        text = font.render_text(msg)
+        assert SDL_GetError() == b""
+        assert isinstance(text, surface.SDL_Surface)
+        assert text.format.contents.format == pixels.SDL_PIXELFORMAT_ARGB8888
+
+        # Test multiline rendering
+        msg = "hello\nthere!"
+        text2 = font.render_text(msg)
+        assert SDL_GetError() == b""
+        assert isinstance(text2, surface.SDL_Surface)
+        assert text2.h > text.h
+
+        # Test custom line height
+        msg = "hello\nthere!"
+        text3 = font.render_text(msg, line_h=100)
+        assert SDL_GetError() == b""
+        assert isinstance(text3, surface.SDL_Surface)
+        assert text3.h > text2.h
+        surface.SDL_FreeSurface(text3)
+
+        # Test wrap width
+        msg = "hello there! This is a very long line of text."
+        text3 = font.render_text(msg, width=200)
+        assert SDL_GetError() == b""
+        assert isinstance(text3, surface.SDL_Surface)
+        assert text3.h > text.h
+        surface.SDL_FreeSurface(text3)
+
+        # Test custom font style (no bg color)
+        msg = "hello there!"
+        font.add_style("blue", 30, (0, 0, 255, 255))
+        text3 = font.render_text(msg, "blue")
+        assert SDL_GetError() == b""
+        assert isinstance(text3, surface.SDL_Surface)
+        surface.SDL_FreeSurface(text3)
+
+        # Test custom font style (w/ bg color)
+        msg = "hello there!"
+        font.add_style("red_on_white", 30, RED_RGBA, (255, 255, 255))
+        text3 = font.render_text(msg, "red_on_white")
+        assert SDL_GetError() == b""
+        assert isinstance(text3, surface.SDL_Surface)
+        surface.SDL_FreeSurface(text3)
+
+        # Test font alignment (not extensive)
+        msg = "hello there!\nThis is a very long line of\ntext."
+        for align in ["left", "center", "right"]:
+            tmp = font.render_text(msg, align=align)
+            assert SDL_GetError() == b""
+            assert isinstance(text3, surface.SDL_Surface)
+            surface.SDL_FreeSurface(tmp)
+
+        surface.SDL_FreeSurface(text)
+        surface.SDL_FreeSurface(text2)
+
+        # Test exception for empty string
+        with pytest.raises(ValueError):
+            font.render_text("")
+
+        # Test exception for missing style
+        with pytest.raises(ValueError):
+            font.render_text(msg, "small")
+
+        # Test exceptions for bad line height
+        with pytest.raises(ValueError):
+            font.render_text(msg, line_h=12.4)
+        with pytest.raises(ValueError):
+            font.render_text(msg, line_h=0)
+
+        # Test exception for bad alignment
+        with pytest.raises(ValueError):
+            font.render_text(msg, align="flush")
+
+    def test_close(self, with_font_ttf):
+        font = with_font_ttf
+        font.close()
+
+        # Test that you can close the font multiple times
+        font.close()
+
+        # Make sure you can't use font after closing it
+        with pytest.raises(RuntimeError):
+            font.get_ttf_font()
+        with pytest.raises(RuntimeError):
+            font.add_style('large', 40, RED_RGBA)
+        with pytest.raises(RuntimeError):
+            font.render_text("hello!")
+        with pytest.raises(RuntimeError):
+            font.contains("A")
+        with pytest.raises(RuntimeError):
+            font.family_name
+        with pytest.raises(RuntimeError):
+            font.style_name
+        with pytest.raises(RuntimeError):
+            font.is_fixed_width
+
+    def test_contains(self, with_font_ttf):
+        font = with_font_ttf
+        for ch in "abcde12345":
+            assert font.contains(ch)
+        for ch in u"的是不":
+            assert not font.contains(ch)
+
+    def test_family_name(self, with_font_ttf):
+        font = with_font_ttf
+        name = font.family_name
+        assert name == None or isinstance(name, str)
+
+    def test_style_name(self, with_font_ttf):
+        font = with_font_ttf
+        name = font.style_name
+        assert name == None or isinstance(name, str)
+
+    def test_is_fixed_width(self, with_font_ttf):
+        font = with_font_ttf
+        assert font.is_fixed_width == False
+
+
+@pytest.mark.skipif(not _HASSDLTTF, reason="SDL_TTF library not available")
+class TestFontManager(object):
+    __tags__ = ["sdl", "sdl2ext"]
+
+    @classmethod
+    def setup_class(cls):
+        try:
+            sdl2ext.init()
+        except sdl2ext.SDLError:
+            raise pytest.skip('Video subsystem not supported')
+
+    @classmethod
+    def teardown_class(cls):
+        sdl2ext.quit()
+
+    def test_init(self):
         fm = sdl2ext.FontManager(RESOURCES.get_path("tuffy.ttf"),
                                  bg_color=(100, 0, 0))
         assert isinstance(fm, sdl2ext.FontManager)
@@ -217,8 +449,7 @@ class TestSDL2ExtFont(object):
         assert fm.size == 16
         assert fm.bg_color == sdl2ext.Color(100, 0, 0, 0)
 
-    @pytest.mark.skipif(not _HASSDLTTF, reason="SDL_TTF library not available")
-    def test_FontManager_default_font(self):
+    def test_default_font(self):
         fm = sdl2ext.FontManager(RESOURCES.get_path("tuffy.ttf"))
         assert fm.default_font == "tuffy"
         assert fm.size == 16
@@ -234,8 +465,7 @@ class TestSDL2ExtFont(object):
         assert fm.default_font == "tuffy.copy"
         assert fm.size == 16
 
-    @pytest.mark.skipif(not _HASSDLTTF, reason="SDL_TTF library not available")
-    def test_FontManager_add(self):
+    def test_add(self):
         fm = sdl2ext.FontManager(RESOURCES.get_path("tuffy.ttf"))
         assert "tuffy" in fm.aliases
         assert "tuffy" in fm.fonts
@@ -262,8 +492,7 @@ class TestSDL2ExtFont(object):
         fm.add(RESOURCES.get_path("tuffy.ttf"), size=12)
         assert isinstance(fm.fonts["tuffy"][12].contents, sdlttf.TTF_Font)
 
-    @pytest.mark.skipif(not _HASSDLTTF, reason="SDL_TTF library not available")
-    def test_FontManager_close(self):
+    def test_close(self):
         fm = sdl2ext.FontManager(RESOURCES.get_path("tuffy.ttf"))
         fm.add(RESOURCES.get_path("tuffy.ttf"), size=20)
         fm.add(RESOURCES.get_path("tuffy.ttf"), alias="Foo", size=10)
@@ -271,8 +500,7 @@ class TestSDL2ExtFont(object):
         assert fm.fonts == {}
         # How to make sure TTF_CloseFont was called on each loaded font?
 
-    @pytest.mark.skipif(not _HASSDLTTF, reason="SDL_TTF library not available")
-    def test_FontManager_render(self):
+    def test_render(self):
         fm = sdl2ext.FontManager(RESOURCES.get_path("tuffy.ttf"))
         text_surf = fm.render("text")
         assert isinstance(text_surf, surface.SDL_Surface)